Subtle bug: shift result is "undefined if the right operand is negative, or greater than *or equal* to the length in bits of the promoted left operand".

+++ include/MultiPrecision/Unsigned.h
@@ -16,7 +16,7 @@ namespace MultiPrecision {
 class Unsigned
 {
 public:
-	using DigitType = unsigned short;
+	using DigitType = unsigned int;
 	static_assert(std::is_integral<DigitType>::value);
 	static_assert(std::is_unsigned<DigitType>::value);
